Made the delay more variable with numbers after the .

+++ minecraft/com/masterof13fps/manager/notificationmanager/Notification.java
@@ -22,7 +22,7 @@ public class Notification {
     private long end;
 
 
-    public Notification(NotificationType type, String title, String messsage, int length) {
+    public Notification(NotificationType type, String title, String messsage, long length) {
         this.type = type;
         this.title = title;
         this.messsage = messsage;
